model.jl

Objects and functions related to reflectometry models.

Jeff.Model — Type

Model(scale::Jeff.Parmeter, bkg::Jeff.Parameter, layers::Array{Jeff.Layer})

The model from which the reflectometry should be calculated.

Parameters

scale::Jeff.Parameter : scale factor for reflectometry.

bkg::Jeff.Parameter : uniform background.

layers::Array{Jeff.Layer} : the array of [Jeff.Layer] objects that describe the model.

Jeff.Parameter — Type

Parameter(value::Float64, vary::Bool, prior::Distrbutions.UnivariateDistribution)

A parameter that can be optimised in the analysis procress.

Parameters

value::Float64 or Distributions.UnivariateDistribution : the value for the given parameter.

Jeff.Layer — Type

Layer(thick::Jeff.Parameter, sld::Jeff.Parameter, isld::Jeff.Parameter, rough::Jeff.Parameter)

A description of a layer in a system.

Parameters

thick::Jeff.Parameter : the layer thickness.

sld::Jeff.Parameter : layer real scattering length density.

isld::Jeff.Parameter : layer imaginary scattering length density.

rough::Jeff.Parameter : roughness with layer above.

Jeff.layers_to_array — Method

layers_to_array(layers::Array{Jeff.Layer})

Convert from an array of N Jeff.Layer objects to an Nx4 array.

Parameters

layers::Array{Jeff.Layer}: layers to be converted.

Returns

::Array{Any, 2} : an array describing the layers.