objective.jl

The code here is focused on optimisation and sampling

Jeff.t_model — Method

t_model(q::Array{Float64, 1}, R::Array{Measurement, 1}, array::Array{Any, 2}, scale::Any, bkg::Any, resolution::Any)

The Turing model for the sampling of the reflectometry profile.

Parameters

q::Array{Float64} : the q-vector values.

R::Array{Measurement} : the reflected intensity values, including uncertainty.

array::Array{Any, 2} : the layers from which the reflectometry should be found.

scale::Any : the scale factor.

bkg::Any : the background level.

resolution::Any : the constant resolution width.