reflect.jl

These functions are focused on the calculation of a reflectometry profile from an abstract set of layers.

Missing docstring.

Missing docstring for Jeff.arbitrary_smearing(q::Array{Float64, 1}, w::Array{Float64, 2}, resolution; nsamples::Int64=500, scale::Float64=1., bkg::Float64=0.). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Jeff.constant_smearing(q::Array{Float64, 1}, w::Array{Float64, 2}; resolution::Float64=0.5, scale::Float64=1., bkg::Float64=0.). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Jeff.abeles(q::Array{Float64, 1}, layers::Array{Float64, 2}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Jeff.same_convolution(a::Array{Float64, 1}, b::Array{Float64, 1}). Check Documenter's build log for details.

Jeff.gauss — Method

gauss(x::Array{Float64, 1}, s::Float64)

A Gaussian kernel for resolution smearing.

Parameters

x::Array{Float64, 1} : the kernal positions.

s::Float64 : the width of the kernel.

Returns

::Array{Float64, 1}: probabilities for x.